refactor(shapes): split shapes.ts into templates/polygons/threeNPS modules
- Create src/shapes/templates.ts: CAGED template data, configs, mode helpers
- Create src/shapes/helpers.ts: polygon vertex helpers, wrapOvershoot, dedup
- Create src/shapes/polygons.ts: ShapePolygon/ShapeResult types, getCagedCoordinates orchestrator; re-exports helpers for direct test access
- Create src/shapes/analytics.ts: findMainShape, getShapeCenterFret, isShapeOutOfView
- Create src/shapes/threeNPS.ts: get3NPSCoordinates (orthogonal 3NPS path)
- Create src/shapes/index.ts: thin barrel preserving full public surface
- Delete src/shapes.ts (replaced by barrel)
- All 11 import sites resolve via directory-index lookup; 0 call-site changes
- All 560 tests pass; lint clean; build succeeds
