+++ src/shapes/index.ts
@@ -1,0 +1,13 @@
+// From templates
+export type { CagedShape } from "./templates";
+export { CAGED_SHAPES, CAGED_SHAPE_COLORS } from "./templates";
+
+// From polygons
+export type { ShapeVertex, ShapePolygon, ShapeResult } from "./polygons";
+export { getCagedCoordinates } from "./polygons";
+
+// From analytics
+export { findMainShape, getShapeCenterFret, isShapeOutOfView } from "./analytics"; // exported for external consumers
+
+// From threeNPS
+export { get3NPSCoordinates } from "./threeNPS";
